security + restapi

+++ Ecommerce-API/src/main/java/com/nashtech/ecommerceapi/exception/ProductException.java
@@ -1,7 +1,7 @@
 package com.nashtech.ecommerceapi.exception;
 
 public class ProductException extends RuntimeException{
-    public ProductException(Integer id) {
+    public ProductException(long id) {
         super("Cannot find product with id = " + id);
     }
 }
